returns array of destinations
