+++ supplemental-api-function-app/build.gradle
@@ -60,8 +60,8 @@ azurefunctions {
     appName = System.getProperty("appName")
     allowTelemetry = false
 
-    auth {
-        type = 'azure_cli'
+    runtime {
+        os = 'windows'
     }
 }
