+++ supplemental-api-function-app/gradle.properties
@@ -1,0 +1,3 @@
+systemProp.subscription=OCIO-EDEDEV-C1
+systemProp.resourceGroup=ocio-ede-dev-moderate-hl7-rg
+systemProp.appName=ocio-ede-dev-supplemental-api
